+++ src/helpers/MediaHelpers.ts
@@ -109,7 +109,11 @@ export class MediaHelpers {
 
       allMedia = [...media];
     } else {
-      if (staticFolder && staticFolder !== STATIC_FOLDER_PLACEHOLDER.hexo.placeholder) {
+      if (
+        staticFolder &&
+        staticFolder !== STATIC_FOLDER_PLACEHOLDER.hexo.placeholder &&
+        staticFolder !== STATIC_FOLDER_PLACEHOLDER.astro.placeholder
+      ) {
         const folderSearch = join(staticFolder || '', '/*');
         const files = await workspace.findFiles(folderSearch);
         const media = await MediaHelpers.updateMediaData(MediaHelpers.filterMedia(files));
@@ -117,6 +121,12 @@ export class MediaHelpers {
         allMedia = [...media];
       } else if (staticFolder && staticFolder === STATIC_FOLDER_PLACEHOLDER.hexo.placeholder) {
         const folderSearch = join(STATIC_FOLDER_PLACEHOLDER.hexo.postsFolder, '/*');
+        const files = await workspace.findFiles(folderSearch);
+        const media = await MediaHelpers.updateMediaData(MediaHelpers.filterMedia(files));
+
+        allMedia = [...media];
+      } else if (staticFolder && staticFolder === STATIC_FOLDER_PLACEHOLDER.astro.placeholder) {
+        const folderSearch = join(STATIC_FOLDER_PLACEHOLDER.astro.assetsFolder, '/*');
         const files = await workspace.findFiles(folderSearch);
         const media = await MediaHelpers.updateMediaData(MediaHelpers.filterMedia(files));
 
@@ -224,6 +234,11 @@ export class MediaHelpers {
       staticPath = join(
         parseWinPath(wsFolder?.fsPath || ''),
         STATIC_FOLDER_PLACEHOLDER.hexo.postsFolder
+      );
+    } else if (staticFolder === STATIC_FOLDER_PLACEHOLDER.astro.placeholder) {
+      staticPath = join(
+        parseWinPath(wsFolder?.fsPath || ''),
+        STATIC_FOLDER_PLACEHOLDER.astro.assetsFolder
       );
     }