+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 - [#424](https://github.com/estruyf/vscode-front-matter/issues/424): Snippet wrapping to allow easier updates or changes to previously set snippets in the content
 - [#585](https://github.com/estruyf/vscode-front-matter/issues/585): New content relationship field type (`contentRelationship`)
 - [#598](https://github.com/estruyf/vscode-front-matter/issues/598): Multilingual support
+- [#615](https://github.com/estruyf/vscode-front-matter/issues/615): Added support for `astro:assets` - [Astro Assets](https://docs.astro.build/en/guides/assets/)
 
 ### 🎨 Enhancements